Convert from remote functions to server load functions
Replaced experimental remote functions with standard SvelteKit server
load functions for better Cloudflare Workers compatibility.
