+++ src/routes/notes/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import NoteLink from './NoteLink.svelte';
-	import { getNotes } from './notes.remote';
+
+	let { data } = $props();
 </script>
 
 <svelte:head>
@@ -9,7 +10,7 @@
 
 <article id="notes" class="prose flow-xs">
 	<ul class="flow" role="list">
-		{#each await getNotes() as post (post.slug)}
+		{#each data.notes as post (post.slug)}
 			<NoteLink {post} />
 		{/each}
 	</ul>
