+++ src/routes/notes/[slug]/+page.server.ts
@@ -1,4 +1,3 @@
-import { query } from '$app/server';
 import { error } from '@sveltejs/kit';
 import matter, { type Input } from 'gray-matter';
 import { marked } from 'marked';
@@ -10,16 +9,6 @@ import css from 'highlight.js/lib/languages/css';
 import bash from 'highlight.js/lib/languages/bash';
 import json from 'highlight.js/lib/languages/json';
 import markdown from 'highlight.js/lib/languages/markdown';
-import * as z from 'zod/mini';
-
-export type Note = {
-	slug: string;
-	title: string;
-	description: string;
-	date: string;
-	draft?: boolean;
-	content: string;
-};
 
 // Register languages
 hljs.registerLanguage('javascript', javascript);
@@ -55,22 +44,8 @@ const noteModules = import.meta.glob('/content/notes/*.md', {
 	import: 'default'
 });
 
-// Parse once at module scope
-const allNotes = Object.entries(noteModules).map(([path, content]) => {
-	const fm = matter(content as Input);
-	const { data, content: md } = fm;
-	const slug = path.match(/\/([^/]+)\.md$/)?.[1];
-	return { slug, ...data, content: md };
-}) as Note[];
-
-// Get all posts
-export const getNotes = query(async () => {
-	return allNotes.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime());
-});
-
-// Get single post by slug
-export const getNote = query(z.string(), async (slug) => {
-	const path = `/content/notes/${slug}.md`;
+export async function load({ params }) {
+	const path = `/content/notes/${params.slug}.md`;
 	const content = noteModules[path];
 
 	if (!content) error(404, 'Post not found');
@@ -78,8 +53,8 @@ export const getNote = query(z.string(), async (slug) => {
 	const { data, content: markdown } = matter(content as Input);
 
 	return {
-		slug,
+		slug: params.slug,
 		meta: data,
 		content: marked(markdown)
 	};
-});
+}
