+++ src/routes/notes/[slug]/+page.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
-	import { page } from '$app/state';
-	import { getNote } from '../notes.remote';
-
-	let param = $derived(page.params.slug);
-	let note = $derived(await getNote(param!));
+	let { data } = $props();
 </script>
 
-<h1>{note.meta.title}</h1>
+<h1>{data.meta.title}</h1>
 
-{@html note.content}
+{@html data.content}
